Fix issue #158: Continue modularizing the orchestration runner after state extraction
